pyproject.toml
[project]
name = "medical-entity-code-mapper"
version = "1.0.0"
description = "High-performance medical entity extraction and coding system"
authors = [{name = "William Dailey, MD, MSEng, MSMI", email = "docdailey@gmail.com"}]
requires-python = ">=3.13,<3.14"
dependencies = [
    "torch==2.7.1",
    "transformers>=4.30.0",
    "sentence-transformers>=2.2.2",
    "faiss-cpu>=1.7.4",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "pandas>=2.0.0",
    "tqdm>=4.65.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
mecm-download-models = "scripts.download_models:main"
mecm-start = "start:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# Let UV use default PyTorch builds for macOS
# macOS ARM64 requires specific builds from PyPI

[tool.uv]
dev-dependencies = [
    "pytest>=7.3.0",
    "black>=23.3.0",
    "flake8>=6.0.0",
    "mypy>=1.3.0",
]